requirements.txt
asgiref==3.4.1
autopep8==2.0.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.2.6
django-cors-headers==3.10.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
idna==3.4
Pillow==9.0.0
pipreqs==0.4.11
pycodestyle==2.9.1
PyJWT==1.7.1
pytz==2021.1
requests==2.28.2
sqlparse==0.4.1
tomli==2.0.1
urllib3==1.26.14
yarg==0.1.9
